Add Generic Error Handling
